Fix type conversion warnings.

+++ drw_entities.h
@@ -965,7 +965,7 @@ public:
     }
 
     void update() {
-        numedges = objlist.size();
+        numedges = (int)objlist.size();
     }
 
 public:
